// vite-movies/src/devils-mode/types/other.ts
export interface Other {
  posts: Post[];
}

export interface Post {
  post_id: string;
  post_title: string;
  post_content: string;
  post_thumbnail: string;
  post_actor: string;
  post_quality: string;
  cate_id: string;
  post_tag: string;
  post_stream: string;
  post_code_id: string;
  post_stream_type: string;
  post_video_code: string;
  source: string;
  create_time: string;
  cate_title: string;
}